Refactored Filesystem : Utilizing .gitignore template : Modified README : Abstracted Verification of cmd args on a function
